# Optimize getRowsColumnRange