--- a/atlasdb-impl-shared/src/main/java/com/palantir/atlasdb/transaction/impl/SnapshotTransaction.java
+++ b/atlasdb-impl-shared/src/main/java/com/palantir/atlasdb/transaction/impl/SnapshotTransaction.java
@@ -489,7 +489,7 @@ public class SnapshotTransaction extends AbstractTransaction implements Constrai
                             raw,
                             x -> x));
             return postFiltered.entrySet().iterator();
-        });
+        }));
     }
 
     private Iterator<Map.Entry<Cell, byte[]>> getRowColumnRangePostFiltered(