--- a/changelog/@unreleased/pr-4668.v2.yml
+++ b/changelog/@unreleased/pr-4668.v2.yml
@@ -1,0 +1,5 @@
+type: improvement
+improvement:
+  description: We expect the performance of `getRowsColumnRange` to have been improved.
+  links:
+  - https://github.com/palantir/atlasdb/pull/4668