We're seeing perf issues internally on the random getRowsColumnRange
implementation that works better on Postgres and Oracle. This is the way
to make it both simpler, and faster.

This is an alternative to https://github.com/palantir/atlasdb/pull/4326.

Needs tests.