@jeremyk-91 2020-03-26
**Postgres**

Old method

```
Benchmark                                                                                                                                                                           (uri)    Mode  Cnt        Score        Error  Units
TransactionGetRowsColumnRangeBenchmarks.getAllColumnsModeratelyWideRow                     com.palantir.atlasdb.performance.backend.PostgresKeyValueServiceInstrumentation@127.0.0.1:5432  sample   42   360647.729 ±  16119.913  us/op
TransactionGetRowsColumnRangeBenchmarks.getAllColumnsModeratelyWideRowWithManyUncommitted  com.palantir.atlasdb.performance.backend.PostgresKeyValueServiceInstrumentation@127.0.0.1:5432  sample   35   437555.785 ±  24811.503  us/op
TransactionGetRowsColumnRangeBenchmarks.getAllColumnsModeratelyWideRowWithSomeUncommitted  com.palantir.atlasdb.performance.backend.PostgresKeyValueServiceInstrumentation@127.0.0.1:5432  sample   37   414031.651 ±  13220.436  us/op
TransactionGetRowsColumnRangeBenchmarks.getAllColumnsSingleBigRow                          com.palantir.atlasdb.performance.backend.PostgresKeyValueServiceInstrumentation@127.0.0.1:5432  sample   28  5900786.542 ± 165164.916  us/op
```

New method

```
Benchmark                                                                                                                                                                           (uri)    Mode  Cnt        Score         Error  Units
TransactionGetRowsColumnRangeBenchmarks.getAllColumnsModeratelyWideRow                     com.palantir.atlasdb.performance.backend.PostgresKeyValueServiceInstrumentation@127.0.0.1:5432  sample   43   354382.110 ±   10197.398  us/op
TransactionGetRowsColumnRangeBenchmarks.getAllColumnsModeratelyWideRowWithManyUncommitted  com.palantir.atlasdb.performance.backend.PostgresKeyValueServiceInstrumentation@127.0.0.1:5432  sample   34   446215.349 ±   44773.869  us/op
TransactionGetRowsColumnRangeBenchmarks.getAllColumnsModeratelyWideRowWithSomeUncommitted  com.palantir.atlasdb.performance.backend.PostgresKeyValueServiceInstrumentation@127.0.0.1:5432  sample   37   415377.795 ±   16679.019  us/op
TransactionGetRowsColumnRangeBenchmarks.getAllColumnsSingleBigRow                          com.palantir.atlasdb.performance.backend.PostgresKeyValueServiceInstrumentation@127.0.0.1:5432  sample   25  6573984.317 ± 2170410.658  us/op
```

I had difficulty getting the Cassandra benchmarks to work properly (some issues with MGMS). Is the perf improvement expected more in that and/or in some case our benchmarks don't test?