@mswintermeyer 2020-03-27
The hope was to improve perf for a situation when calling getRowsColumnRange across many rows, but where each have just ~1 column each. Previously we'd fallback to the BatchSizeIncreasingIterator which would throw away the original iterator and create a new one regularly, and each time creating a new one would mean validating the locks are still valid.