@jeremyk-91 2020-03-27
Cool. I added a test that features a long and narrow table with 1 million rows and 1 column. Also added a 2 milli penalty for a lock check to simulate using timelock. Then,

new
```
Benchmark                                                                                                                                    (uri)    Mode  Cnt       Score      Error  Units
TransactionGetRowsColumnRangeBenchmarks.jeremyTest  com.palantir.atlasdb.performance.backend.PostgresKeyValueServiceInstrumentation@127.0.0.1:5432  sample  804  111896.851 ± 4578.068  us/op
```

old
```
Benchmark                                                                                                                                    (uri)    Mode  Cnt         Score         Error  Units
TransactionGetRowsColumnRangeBenchmarks.jeremyTest  com.palantir.atlasdb.performance.backend.PostgresKeyValueServiceInstrumentation@127.0.0.1:5432  sample    4  23412604.928 ± 3275831.242  us/op
```