@jeremyk-91 2020-03-27
With 2 ms penalty for timelock ops:

new
```
TransactionGetRowsColumnRangeBenchmarks.getAllColumnsModeratelyWideRow                     com.palantir.atlasdb.performance.backend.PostgresKeyValueServiceInstrumentation@127.0.0.1:5432  sample   40   380003.942 ±  14246.615  us/op
TransactionGetRowsColumnRangeBenchmarks.getAllColumnsModeratelyWideRowWithManyUncommitted  com.palantir.atlasdb.performance.backend.PostgresKeyValueServiceInstrumentation@127.0.0.1:5432  sample   34   451134.404 ±  15757.661  us/op
TransactionGetRowsColumnRangeBenchmarks.getAllColumnsModeratelyWideRowWithSomeUncommitted  com.palantir.atlasdb.performance.backend.PostgresKeyValueServiceInstrumentation@127.0.0.1:5432  sample   35   437810.439 ±  18242.531  us/op
TransactionGetRowsColumnRangeBenchmarks.getAllColumnsSingleBigRow                          com.palantir.atlasdb.performance.backend.PostgresKeyValueServiceInstrumentation@127.0.0.1:5432  sample   27  6049739.814 ± 175294.550  us/op

@mswintermeyer I think all of these benchmarks only have one row, though, so the improvement in terms of reducing the number of lock checks would not be significant.
```

old
```
Benchmark                                                                                                                                                                           (uri)    Mode  Cnt        Score        Error  Units
TransactionGetRowsColumnRangeBenchmarks.getAllColumnsModeratelyWideRow                     com.palantir.atlasdb.performance.backend.PostgresKeyValueServiceInstrumentation@127.0.0.1:5432  sample   40   375875.174 ±  17092.756  us/op
TransactionGetRowsColumnRangeBenchmarks.getAllColumnsModeratelyWideRowWithManyUncommitted  com.palantir.atlasdb.performance.backend.PostgresKeyValueServiceInstrumentation@127.0.0.1:5432  sample   33   459308.063 ±  27465.031  us/op
TransactionGetRowsColumnRangeBenchmarks.getAllColumnsModeratelyWideRowWithSomeUncommitted  com.palantir.atlasdb.performance.backend.PostgresKeyValueServiceInstrumentation@127.0.0.1:5432  sample   35   432088.210 ±  14429.160  us/op
TransactionGetRowsColumnRangeBenchmarks.getAllColumnsSingleBigRow                          com.palantir.atlasdb.performance.backend.PostgresKeyValueServiceInstrumentation@127.0.0.1:5432  sample   26  6184339.929 ± 194572.677  us/op
```