@svc-autorelease 2020-03-27
Released [0.198.9](https://github.com/palantir/atlasdb/releases/tag/0.198.9)